Add drain to black mage
